Task 1 ' Logo creation'
I found out that the Project deals with Artificial Intelligence coupled with human Clinical data to create a better diagnostic tool for typhoid fever. So I came up with the idea to include, an AI brain and a telescope to showcase the findings of better diagnostics for typhoid fever. I tried making sure that with a single glance at the logo, one would be able to tell what it's all about.