Task 3: Develope a data collection tool for symptom data
I designed and developed an android app called iMDOTA Survey which is an offline data collection tool. It'll help especially those in the remote areas.
This tool (android app) collect the associations of symptoms that confirm the disease and its level of severity.

Main Idea
This data collection tool (android app) is in the form of survey, It collect typhoid symptoms and send them to an online platform for data analysis, This online platform can be a database management system or anything that processes data and store it so it can be reviewed by experts.

RECOMMENDED DATA ELEMENTS (I used them for the survey)
In systems where case-based data are collected for
reporting and investigation, the following parameters are
recommended for collection. It is important to ensure
that clinical data are linked to laboratory data for each
case.

This data collection tool is going to help collect offline data in a structured manner

    Unique identifier

    Date of report

    Age/date of birth

    Sex

    Country of birth

    Place of residence (city, district and province)

    Travel history within the last 28 days before illness

    onset for persons not living in endemic settings

    (places and dates)

    Household contact with a confirmed case of typhoid or

    paratyphoid fever in the 28 days before illness onset

    Fever for at least three out of seven consecutive days

    Date of fever onset

    Are You Having Fever?

    What is your current temperature?

    What are your symptoms?
    -headache
    -weakness and fatigue
    -Muscke aches
    -sweating
    -dry cough
    -loss of appetite and weigh loss
    -stomach pain

    Hospitalization (including date of admission and

    date of discharge)

    Complications (intestinal perforation,

    encephalopathy)

    Abdominal surgery

    Outcome/discharge status

    Type of specimen(s) collected

    Date of specimen(s) collection

    Date specimen(s) received in laboratory

App Screenshots