Task 1 Designing the Project logo
Task 1 Designing a Logo for the Project
Designing a Logo for the Project

My logo focuses on th main contributors to the project Mboalab and shows it written in the rich Cameroonian/African lines and culture
The project is on the study of Typhoid Fever, which is caused by the Salmonella Typhi, this is represented by the drawings to the left of the logo, seen under the M for Mboalab. 
The project is going to employ the use of Artificial Intelligence and tis isrepresented by the original drawing of the face to the right of the logo. 
It is drawn with its tentacles reaching out to the Salmonella Typhi. This is to depict an engulfing of the data on Salmonella Typhi to gain insight that will be of benefit to the immediate environment and subsequently the whole world.

My logo was designed using GIMP, an open source graphic design tool.